- (snmpv3.c, snmpv3.h): use size_t for engineid lengths.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@4094 06827809-a52a-0410-b366-d66718629ded

+++ snmplib/snmpv3.h
@@ -34,8 +34,8 @@ int     snmpv3_store(int majorID, int minorID, void *serverarg,
                      void *clientarg);
 u_long  snmpv3_local_snmpEngineBoots(void);
 int     snmpv3_clone_engineID(u_char **, size_t* , u_char*, size_t);
-u_long  snmpv3_get_engineID(u_char *buf, size_t buflen);
-u_char *snmpv3_generate_engineID(u_long *);
+size_t  snmpv3_get_engineID(u_char *buf, size_t buflen);
+u_char *snmpv3_generate_engineID(size_t *);
 u_long  snmpv3_local_snmpEngineTime(void);
 int     get_default_secLevel(void);
 oid    *get_default_authtype(size_t *);
